Created common book names
Book names are set to a standard name at the initial parsing of the query using the biblejs-name-converter. This was previously just called later in the data pipeline for the bolly bible provider, but by moving it to the beginning of parsing, it makes the reference head, backlinks, and tags all utilize the same string.